fix(knockdog): dynamic import를 사용하여 client component를 가져올 때 ssr 옵션 비활성화하도록 변경
@knockdog/ui/BottomSheet 내부에서 vaul 라이브러리를 사용
vaul 라이브러리에서 ssr 환경에 대한 고려가 없음(document is not defined 에러) → 차선책으로 dynamic
import ssr 옵션 비활성화로 설정
